+++ tests/test_validator.py
@@ -1,83 +1,68 @@
-from baby_steps import then, when
-from district42 import schema
+from unittest.mock import Mock, call, sentinel
+
+from baby_steps import given, then, when
+from district42 import Props
+from district42.types import Schema
 from pytest import raises
+from th import PathHolder
 
-from valera import ValidationException, ValidationResult, eq, validate, validate_or_fail
+from valera import ValidationResult, Validator
 
 
-def test_validator_validate_pass():
+def test_validator_path_holder_factory():
+    with given:
+        validator = Validator()
+
     with when:
-        result = validate(schema.int, 42)
+        path = validator.make_path()
 
     with then:
-        assert isinstance(result, ValidationResult)
-        assert not result.has_errors()
+        assert isinstance(path, PathHolder)
 
 
-def test_validator_validate_fail():
+def test_validator_validation_result_factory():
+    with given:
+        validator = Validator()
+
     with when:
-        result = validate(schema.int, "42")
+        path = validator.make_validation_result()
 
     with then:
-        assert isinstance(result, ValidationResult)
-        assert result.has_errors()
+        assert isinstance(path, ValidationResult)
 
 
-def test_validator_validate_or_fail():
+def test_validator_visit():
+    with given:
+        mock = Mock(return_value=sentinel.visited)
+
+        class CustomType(Schema[Props]):
+            def __valera__(self, *args, **kwargs) -> str:
+                return mock(*args, **kwargs)
+
+        custom_type = CustomType()
+        validator = Validator()
+
     with when:
-        result = validate_or_fail(schema.int, 42)
+        res = validator.visit(custom_type, value=sentinel.value, path=sentinel.path)
 
     with then:
-        assert result is True
+        assert res is sentinel.visited
+        assert mock.mock_calls == [
+            call(validator, value=sentinel.value, path=sentinel.path)
+        ]
 
 
-def test_validator_validate_or_fail_error():
+def test_validator_visit_error():
+    with given:
+        class CustomType(Schema[Props]):
+            pass
+
+        custom_type = CustomType()
+        validator = Validator()
+
     with when, raises(Exception) as exception:
-        validate_or_fail(schema.dict, "")
+        validator.visit(custom_type)
 
     with then:
-        assert exception.type is ValidationException
-        assert str(exception.value) == ("\n - Value '' must be <class 'dict'>, "
-                                        "but <class 'str'> given")
-
-
-def test_validator_validate_or_fail_multiline():
-    with when, raises(Exception) as exception:
-        validate_or_fail(schema.dict({"id": schema.int, "name": schema.str}), {})
-
-    with then:
-        assert exception.type is ValidationException
-        assert str(exception.value) == ("\n - Key _['id'] does not exist"
-                                        "\n - Key _['name'] does not exist")
-
-
-def test_validator_eq():
-    with when:
-        result = eq(schema.str, schema.str)
-
-    with then:
-        assert result is True
-
-
-def test_validator_not_eq_schema():
-    with when:
-        result = eq(schema.str, schema.int)
-
-    with then:
-        assert result is False
-
-
-def test_validator_not_eq_props():
-    with when:
-        result = eq(schema.str, schema.str.len(1))
-
-    with then:
-        assert result is False
-
-
-def test_validator_not_eq_value():
-    with when:
-        result = eq(schema.str, 42)
-
-    with then:
-        assert result is False
+        assert exception.type is NotImplementedError
+        assert str(exception.value) == "CustomType has no method '__valera__'"